@@ -19,12 +19,12 @@ interface Authorization<T> {
         user: String? = null,
         source: String? = null
     ): Boolean {
-        return hasPermission(token, permission, entity, permissionOn, project, user, source)
-                && hasRole(token, project, role)
-                && hasScopes(token, scopes)
-                && hasAuthorities(token, authorities)
-                && hasAudiences(token, audiences)
-                && hasGrantTypes(token, grantTypes)
+        return hasPermission(token, permission, entity, permissionOn, project, user, source) &&
+            hasRole(token, project, role) &&
+            hasScopes(token, scopes) &&
+            hasAuthorities(token, authorities) &&
+            hasAudiences(token, audiences) &&
+            hasGrantTypes(token, grantTypes)
     }
 
     fun hasPermission(
