@@ -5,7 +5,6 @@ import org.springframework.web.bind.annotation.ResponseStatus
 
 @ResponseStatus(HttpStatus.FORBIDDEN)
 class ResourceForbiddenException @JvmOverloads constructor(
-    override val message: String, override val cause: Throwable? =
-        null
-) :
-    RuntimeException(message, cause)
+    override val message: String,
+    override val cause: Throwable? = null
+) : RuntimeException(message, cause)
