@@ -1,5 +1,6 @@
 package radar.spring.auth.managementportal
 
+import jakarta.servlet.http.HttpServletRequest
 import org.radarbase.auth.authentication.TokenValidator
 import org.radarbase.auth.config.TokenValidatorConfig
 import org.radarbase.auth.config.TokenVerifierPublicKeyConfig
@@ -11,7 +12,6 @@ import org.springframework.stereotype.Component
 import radar.spring.auth.common.RadarAuthValidator
 import radar.spring.auth.config.ManagementPortalAuthProperties
 import java.net.URI
-import javax.servlet.http.HttpServletRequest
 
 /** The [radar.spring.auth.common.AuthValidator] for Management Portal tokens. **/
 @Component
